BPM.ENH: remove single pass graphs and move triggered acquisition windows to trig_acq_config module
