+++ pyqt-apps/siriushla/as_di_bpms/__init__.py
@@ -1,5 +1,4 @@
-from .list_bpms import SelectBPMs, SinglePassSummary, MultiTurnSummary, \
-    BPMSummary
+from .list_bpms import SelectBPMs, AcqDataSummary, BPMSummary
 from .main import BPMMain
 from .monit import MonitData
 from .settings import ParamsSettings, \
