+++ pyqt-apps/siriushla/as_di_bpms/list_bpms.py
@@ -81,148 +81,11 @@ class SelectBPMs(BaseWidget):
         self.scarea.verticalScrollBar().setValue(0)
 
 
-class SinglePassSummary(BaseWidget):
-
-    def __init__(self, parent=None, prefix='', bpm_list=[]):
-        super().__init__(
-            parent=parent, prefix=prefix, bpm='', data_prefix='SP_')
-        self.bpm_dict = {bpm: '' for bpm in bpm_list}
-        self._name = bpm_list[0][:2] + 'App'
-        self.setObjectName(self._name)
-        self.setStyleSheet('#'+self._name+'{min-width:65em;min-height:38em;}')
-        self.setupui()
-
-    def setupui(self):
-        vbl = QVBoxLayout(self)
-        lab = QLabel('<h2>BPMs List</h2>', alignment=Qt.AlignCenter)
-        vbl.addWidget(lab)
-        vbl.addSpacing(20)
-
-        hbl = QHBoxLayout()
-        search = QLineEdit(parent=self)
-        search.setPlaceholderText("Search for BPMs...")
-        search.textEdited.connect(self._filter_bpms)
-        hbl.addWidget(search)
-        hbl.addStretch()
-        self.btnautorange = QPushButton('Auto Range graphics', self)
-        hbl.addWidget(self.btnautorange)
-        vbl.addItem(hbl)
-
-        sa_class = get_custom_widget_class(QScrollArea)
-        scarea = sa_class(self)
-        scarea.setSizeAdjustPolicy(scarea.AdjustToContents)
-        scarea.setWidgetResizable(True)
-
-        wid = QWidget()
-        wid.setObjectName('scrollarea')
-        wid.setStyleSheet('#scrollarea {background-color: transparent;}')
-        gdl = QGridLayout(wid)
-        gdl.setSpacing(15)
-        for i, bpm in enumerate(sorted(self.bpm_dict.keys())):
-            widb = QWidget(wid)
-            vbl2 = QVBoxLayout(widb)
-            vbl2.addWidget(QLabel(
-                '<h3>'+bpm+'</h3>', alignment=Qt.AlignCenter))
-            wbpm = self.create_graph(widb, bpm=bpm, typ='ant')
-            vbl2.addWidget(wbpm)
-
-            gdl.addWidget(widb, i // 3, i % 3)
-            self.bpm_dict[bpm] = widb
-        self.gdl = gdl
-        vbl.addWidget(scarea)
-        scarea.setWidget(wid)
-        self.scarea = scarea
-
-    def create_graph(self, wid, bpm, typ='pos'):
-        text, unit, names, colors = self._get_properties(typ)
-        if typ == 'pos':
-            unit = unit[1:]
-            suff = '-Mon'
-            CLASS = GraphTime
-        elif typ == 'ant':
-            suff = 'ArrayData'
-            CLASS = GraphWave
-        else:
-            suff = '-Mon'
-            CLASS = GraphTime
-
-        graph = CLASS(
-            wid, prefix=self.prefix, bpm=bpm,
-            data_prefix=self.data_prefix)
-        graph.maxRedrawRate = 2.1
-        graph.graph.plotItem.vb.autoRange()
-        self.btnautorange.clicked.connect(graph.graph.plotItem.vb.autoRange)
-        graph.setLabel('left', text=text, units=unit)
-        for name, cor in zip(names, colors):
-            opts = dict(
-                y_channel=name+suff,
-                name=name[2:],
-                color=cor,
-                lineStyle=1,
-                lineWidth=1)  # NOTE: If > 1: very low performance
-            if typ == 'pos':
-                opts['y_channel'] = graph.get_pvname(
-                    opts['y_channel'], is_data=False)
-                graph.addYChannel(add_scale=1e-9, **opts)
-            elif typ == 'amp':
-                opts['y_channel'] = graph.get_pvname(
-                    opts['y_channel'], is_data=False)
-                graph.addYChannel(**opts)
-            else:
-                opts['name'] = text[:3] + name
-                opts['y_channel'] = graph.get_pvname(opts['y_channel'])
-                graph.addChannel(**opts)
-        graph.setObjectName('graph')
-        graph.setStyleSheet('#graph{min-width: 18em; min-height: 12em;}')
-        return graph
-
-    def _get_properties(self, typ='pos'):
-        if typ == 'pos':
-            text = 'Positions'
-            unit = 'nm'
-            names = ('SPPosX', 'SPPosY', 'SPPosQ', 'SPSum')
-            colors = ('blue', 'red', 'green', 'black')
-        elif typ == 'ant':
-            text = 'Antennas'
-            unit = 'count'
-            names = ('A', 'B', 'C', 'D')
-            colors = ('blue', 'red', 'green', 'magenta')
-        if typ == 'amp':
-            text = 'Amplitudes'
-            unit = 'count'
-            names = ('SPAmplA', 'SPAmplB', 'SPAmplC', 'SPAmplD')
-            colors = ('blue', 'red', 'green', 'magenta')
-        return text, unit, names, colors
-
-    @Slot(str)
-    def _filter_bpms(self, text):
-        """Filter power supply widgets based on text inserted at line edit."""
-        try:
-            pattern = re.compile(text, re.I)
-        except Exception:
-            return
-
-        for i in range(self.gdl.rowCount()):
-            for j in range(self.gdl.columnCount()):
-                self.gdl.removeItem(self.gdl.itemAtPosition(i, j))
-        wids = []
-        for bpm, wid in self.bpm_dict.items():
-            mat = bool(pattern.search(bpm))
-            wid.setVisible(mat)
-            if mat:
-                wids.append(wid)
-
-        for i, wid in enumerate(wids):
-            self.gdl.addWidget(wid, i // 3, i % 3)
-        # Sroll to top
-        self.scarea.verticalScrollBar().setValue(0)
-
-
-class MultiTurnSummary(BaseWidget):
+class AcqDataSummary(BaseWidget):
 
     def __init__(self, parent=None, prefix='', bpm_list=[], mode='pos'):
         super().__init__(
-            parent=parent, prefix=prefix, bpm='', data_prefix='GEN_')
+            parent=parent, prefix=prefix, bpm='', data_prefix='GEN')
         self.bpm_dict = {bpm: '' for bpm in bpm_list}
         self.mode = mode.lower()
         self._name = bpm_list[0][:2] + 'App'
@@ -285,8 +148,8 @@ class MultiTurnSummary(BaseWidget):
         graph.setLabel('left', text=text, units=unit)
         for name, cor in zip(names, colors):
             opts = dict(
-                y_channel=name+'ArrayData',
-                name=text[:3]+name,
+                y_channel=name+'Data',
+                name=name,
                 color=cor,
                 lineStyle=1,
                 lineWidth=1)  # NOTE: If > 1: very low performance
@@ -303,12 +166,12 @@ class MultiTurnSummary(BaseWidget):
         if typ.startswith('pos'):
             text = 'Positions'
             unit = 'nm'
-            names = ('X', 'Y', 'Q', 'SUM')
+            names = ('PosX', 'PosY', 'PosQ', 'Sum')
             colors = ('blue', 'red', 'green', 'black')
         else:
             text = 'Antennas'
             unit = 'count'
-            names = ('A', 'B', 'C', 'D')
+            names = ('AmplA', 'AmplB', 'AmplC', 'AmplD')
             colors = ('blue', 'red', 'green', 'magenta')
         return text, unit, names, colors
